Refactor Gateway for MAG
This commit add's new globalnetdatapath controllers
to the gateway code which handle programming
all the needed datapath bits on each active
GW node when globalnet is enabled.

For this to function we need to add the iptables
binary to the gateway pod.

It also adds some new bits specifically for the
MAG feature.

Now on all GWs return routes are setup based on
each remote GW SNAT IP list, so that it's possible
to route return traffic back to the originating
remote gateway.

Also, in order for each gateway to be able to have
healthchecks for all remote gateways, routes are setup
to correctly direct traffic to each assigned remote
GW node global IP.

Lastly, in order to ensure that intra cluster remote sourced
svc traffic returns back to the originating local gateway,
a SNAT rule is added to the `SUBMARINER-GN-EGRESS` chain
which SNATs all traffic entering via the `vxlan-tunnel`
interface destined for a local SVC (i.e the weave interface)
to the local weave interface IP.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>